New snippet build
