+++ node/install-snippets.js
