+++ preview/css/_index.css
@@ -2,7 +2,6 @@
 @import "https://fonts.googleapis.com/css?family=Work+Sans:500,700&display=swap"; /* todo - add to eb-fonts */
 @import "normalize.css";
 @import "@ekstra-bladet/eb-colors/dist/eb-colors-classes.css";
-@import "@ekstra-bladet/eb-fonts/dist/eb-fonts-desktop.css";
 @import "_highlight";
 
 html,
@@ -68,12 +67,12 @@ h2 {
 
 @media (max-width: 767px) {
   .dsm-nav-top-desktop .nav-top {
-  padding-right: 10px;
+    padding-right: 10px;
   padding-left: 10px;
   min-width: 300px;
   }
   .nav-vertical {
-  position: fixed;
+    position: fixed;
   z-index: 1;
 
   background: #eee;
